BUG: Fix Memory Leak in DicomServer
When accepting a connection, a cancellation token is passed to allow aborting the connection. However, the token that is passed causes a circular reference in the created task. This in turn causes the task to not be collected by the GC when the task is cleaned up.

Use a linked token source to prevent this leak, as the derived token is disposed after the network stream is accepted and no longer needed.